Split code to separate files.
