+++ lumimqtt/device.py
@@ -1,0 +1,7 @@
+class Device:
+    MQTT_VALUES = None
+
+    def __init__(self, device, name, topic):
+        self.name = name
+        self.device = device
+        self.topic = topic
