+++ lumimqtt/lumimqtt.py
@@ -4,295 +4,22 @@ import logging
 import typing as ty
 from dataclasses import dataclass
 from datetime import datetime
-from os.path import exists
-from subprocess import DEVNULL, CalledProcessError, run
 
 import aio_mqtt
-from evdev import InputDevice, KeyEvent, categorize, ecodes  # noqa
 
 from .__version__ import VERSION
+from .button import Button
+from .device import Device
+from .light import Light
+from .sensors import BinarySensor, Sensor
 
 logger = logging.getLogger(__name__)
-
-
-class ButtonAction:
-    SINGLE = 'single'
-    DOUBLE = 'double'
-    TRIPLE = 'triple'
-    QUADRUPLE = 'quadruple'
-    MANY = 'many'
-    HOLD = 'hold'
-    DOUBLE_HOLD = 'double_hold'
-    TRIPLE_HOLD = 'triple_hold'
-    QUADRUPLE_HOLD = 'quadruple_hold'
-    MANY_HOLD = 'many_hold'
-    RELEASE = 'release'
 
 
 @dataclass
 class DebounceSensor:
     value: ty.Any
     last_sent: datetime
-
-
-class Device:
-    MQTT_VALUES = None
-
-    def __init__(self, device, name, topic):
-        self.name = name
-        self.device = device
-        self.topic = topic
-
-
-class Sensor(Device):
-    def get_value(self):
-        raise NotImplementedError()
-
-
-class Button(Device):
-    MQTT_VALUES = {
-        'icon': 'mdi:gesture-double-tap',
-    }
-    THRESHOLD = 0.3
-    PROVIDE_EVENTS = [
-        getattr(ButtonAction, x)
-        for x in dir(ButtonAction) if not x.startswith('__')
-    ]
-
-    def __init__(self, device, name, topic, scancodes=None):
-        super().__init__(device, name, topic)
-        self.ev_device = InputDevice(self.device)
-        self.scancodes = scancodes
-
-        self.event_queue = None
-        self.is_pressed = False
-        self.is_sent = False
-        self.clicks_done = 0
-
-    async def handle_events(self):
-        async for event in self.ev_device.async_read_loop():
-            event = categorize(event)
-            if isinstance(event, KeyEvent) and (
-                not self.scancodes or event.scancode in self.scancodes
-            ):
-                if event.keystate in [KeyEvent.key_up, KeyEvent.key_down]:
-                    await self.event_queue.put(event.keystate)
-
-    async def handle_queue(self, on_click):
-        while True:
-            if self.is_pressed and not self.is_sent or self.clicks_done:
-                try:
-                    event = await aio.wait_for(
-                        self.event_queue.get(),
-                        timeout=self.THRESHOLD,
-                    )
-                except aio.TimeoutError:
-                    action = {
-                        (False, 1): ButtonAction.SINGLE,
-                        (False, 2): ButtonAction.DOUBLE,
-                        (False, 3): ButtonAction.TRIPLE,
-                        (False, 4): ButtonAction.QUADRUPLE,
-                        (True, 0): ButtonAction.HOLD,
-                        (True, 1): ButtonAction.DOUBLE_HOLD,
-                        (True, 2): ButtonAction.TRIPLE_HOLD,
-                        (True, 3): ButtonAction.QUADRUPLE_HOLD,
-                    }.get((self.is_pressed, self.clicks_done))
-                    if action is None:
-                        if self.clicks_done > 3 and self.is_pressed:
-                            action = ButtonAction.MANY_HOLD
-                        elif self.clicks_done > 4 and not self.is_pressed:
-                            action = ButtonAction.MANY
-                        else:
-                            raise NotImplementedError('Unknown button state')
-
-                    await on_click(self, action)
-                    self.is_sent = self.is_pressed
-                    self.clicks_done = 0
-                else:
-                    if event == KeyEvent.key_up:
-                        self.clicks_done += 1
-                        self.is_pressed = False
-                    elif event == KeyEvent.key_down:
-                        self.is_pressed = True
-            else:
-                event = await self.event_queue.get()
-                if event == KeyEvent.key_up:
-                    self.is_pressed = False
-                    await on_click(self, ButtonAction.RELEASE)
-                elif event == KeyEvent.key_down:
-                    self.is_pressed = True
-                self.is_sent = False
-
-    async def handle(self, on_click):
-        self.event_queue = aio.Queue()
-        await aio.gather(
-            self.handle_events(),
-            self.handle_queue(on_click),
-        )
-
-
-class Light(Device):
-    RGB = True
-    BRIGHTNESS = True
-
-    def __init__(self, device: dict, name, topic):
-        super().__init__(device, name, topic)
-        self.led_r = self.device['r']
-        self.led_g = self.device['g']
-        self.led_b = self.device['b']
-        self.pwm_max = int(self.device['pwm_max'])
-
-        state_r = int(open(self.led_r, 'r').read())
-        state_g = int(open(self.led_g, 'r').read())
-        state_b = int(open(self.led_b, 'r').read())
-
-        self.state = {
-            'state': 'ON' if state_r or state_g or state_b else 'OFF',
-            'brightness': 255,
-            'color': {
-                'r': int(state_r / self.pwm_max * 255),
-                'g': int(state_g / self.pwm_max * 255),
-                'b': int(state_b / self.pwm_max * 255),
-            },
-        }
-
-    @property
-    def topic_set(self):
-        return f'{self.topic}/set'
-
-    async def write(self, value: dict):
-        state = value.get('state', self.state['state'])
-        color = value.get('color', self.state['color'])
-        brightness = value.get('brightness', self.state['brightness'])
-
-        for c, file in [
-            ('r', self.led_r),
-            ('g', self.led_g),
-            ('b', self.led_b),
-        ]:
-            pwm_value = \
-                int((color[c] * self.pwm_max / 255) * brightness / 255)
-            if state.lower() == 'off':
-                pwm_value = 0
-            if not (0 <= pwm_value <= self.pwm_max):
-                pwm_value = 0
-            with open(file, 'w+') as f:
-                f.write(str(pwm_value))
-                f.write('\n')
-
-    async def set(self, value: dict):
-        state = value.get('state', self.state['state'])
-        color = value.get('color', self.state['color'])
-        brightness = value.get('brightness', self.state['brightness'])
-        current_brightness = self.state['brightness']
-        target_brightness = brightness
-        if self.state['state'].lower() == 'off':
-            current_brightness = 0
-            if color['r'] == 0 and color['g'] == 0 and color['b'] == 0:
-                color['r'] = 255
-                color['g'] = 255
-                color['b'] = 255
-        if state.lower() == 'off':
-            target_brightness = 0
-        transition = value.get('transition', 1)  # seconds
-        steps = 12 * round(transition + 0.49)
-
-        if transition:
-            start_level = current_brightness
-
-            """ Use brightness or convert brightness_pct """
-            end_level = int(target_brightness)
-
-            if start_level == end_level:
-                await self.write(value)
-                self.state = {
-                    'state': state,
-                    'brightness': brightness,
-                    'color': color,
-                }
-                return
-
-            """ Calculate number of steps """
-            total_range = abs(start_level - end_level)
-            fadeout = start_level > end_level
-
-            """ Calculate the delay time """
-            step_by = total_range / steps
-            delay = transition / steps / 3
-
-            new_level = start_level
-            for x in range(steps):
-                if fadeout:
-                    new_level = new_level - step_by
-                    if new_level < end_level:
-                        new_level = end_level
-                else:
-                    new_level = new_level + step_by
-                    if new_level > end_level:
-                        new_level = end_level
-                new_value = value.copy()
-                new_value['brightness'] = new_level
-                new_value['state'] = 'ON'
-                await self.write(new_value)
-                await aio.sleep(delay)
-            if new_level != target_brightness:
-                new_value = value.copy()
-                new_value['brightness'] = target_brightness
-                await self.write(new_value)
-        else:
-            await self.write(value)
-
-        self.state = {
-            'state': state,
-            'brightness': brightness,
-            'color': color,
-        }
-
-
-class BinarySensor(Sensor):
-    MQTT_VALUES = {}
-
-    def __init__(self, gpio, name, topic, device_class=None):
-        device = f"/sys/class/gpio/gpio{gpio}/value"
-        super().__init__(device, name, topic)
-        if device_class:
-            self.MQTT_VALUES['device_class'] = device_class
-        if not exists(device):
-            try:
-                run(
-                    ['tee', '/sys/class/gpio/export'],
-                    stdout=DEVNULL,
-                    input=str(gpio).encode(),
-                    check=True,
-                )
-                run(
-                    ['tee', f'/sys/class/gpio/gpio{gpio}/direction'],
-                    stdout=DEVNULL,
-                    input='in'.encode(),
-                    check=True,
-                )
-            except CalledProcessError as err:
-                logger.error(f"Can not setup {name} sensor: {err.stdout}")
-
-    def get_value(self):
-        with open(self.device, 'r') as f:
-            return 'OFF' if f.read()[:-1] == '0' else 'ON'
-
-
-class IlluminanceSensor(Sensor):
-    COEFFICIENT = 0.25
-    MQTT_VALUES = {
-        'device_class': 'illuminance',
-        'unit_of_measurement': 'lx',
-    }
-
-    def get_value(self):
-        with open(self.device, 'r') as f:
-            data = f.read()[:-1]
-            try:
-                return int(int(data) * self.COEFFICIENT)
-            except ValueError:
-                return data
 
 
 class LumiMqtt:
